# microservice/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The gateway of forum",
        "title": "forum-gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "forum.muxixyz.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/sip-score/entry/rating": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sipscore"
                ],
                "summary": "修改评分 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "update_sip_score_entry_rating_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sipscore.UpdateSipScoreEntryRatingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sipscore"
                ],
                "summary": "创建榜单条目评分 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "create_sip_score_entry_rating_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sipscore.CreateSipScoreEntryRatingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sipscore.EmptyResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sipscore"
                ],
                "summary": "删除评分 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "delete_sip_score_entry_rating_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sipscore.DeleteSipScoreEntryRatingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
        "/sip-score/{sip_score_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sipscore"
                ],
                "summary": "获取榜单元数据 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sip_score_id",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sipscore.GetSipScoreResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sipscore"
                ],
                "summary": "删除榜单 api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sip_score_id",
                        "name": "sip_score_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "sipscore.CreateSipScoreEntryRatingRequest": {
            "type": "object",
            "required": [
                "comment",
                "entry_id",
                "rating",
                "sip_score_id"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "entry_id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "sip_score_id": {
                    "type": "integer"
                }
            }
        },
...
        "sipscore.DeleteSipScoreEntryRatingRequest": {
            "type": "object",
            "required": [
                "entry_id",
                "rating_id",
                "sip_score_id"
            ],
            "properties": {
                "entry_id": {
                    "type": "integer"
                },
                "rating_id": {
                    "type": "integer"
                },
                "sip_score_id": {
                    "type": "integer"
                }
            }
        },
        "sipscore.EmptyResponse": {
            "type": "object"
        },
...
        "sipscore.GetSipScoreResponse": {
            "type": "object",
            "properties": {
                "sip_score": {
                    "$ref": "#/definitions/sipscore.SipScore"
                }
            }
        },
...
        "sipscore.SipScore": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "collect_count": {
                    "type": "integer"
                },
                "cover_img": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/sipscore.userInfo"
                },
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "entry_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_collected": {
                    "type": "boolean"
                },
                "last_modified_by": {
                    "$ref": "#/definitions/sipscore.userInfo"
                },
                "name": {
                    "type": "string"
                },
                "participant_count": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "sipscore.UpdateSipScoreEntryRatingRequest": {
            "type": "object",
            "required": [
                "entry_id",
                "rating_id",
                "sip_score_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "entry_id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "rating_id": {
                    "type": "integer"
                },
                "sip_score_id": {
                    "type": "integer"
                }
            }
        },
...
        "sipscore.userInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
...
}